# pos-invoice/openapi.yaml
openapi: 3.1.0
info:
  title: POS Invoice API
  description: API documentation for invoice and billing rules management
  contact:
    name: Durion Support Services
    email: platform@durionpos.org
  version: v1
servers:
- url: http://localhost:8086
  description: Generated server url
...
paths:
...
  /v1/invoices:
    post:
      tags:
      - Invoice
      summary: Create invoice
      description: Create invoice draft from completed workorder data
      operationId: createInvoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCreationRequest'
        required: true
      responses:
        '201':
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceGenerationResponse'
      security:
      - bearerAuth:
        - invoice:manage
      x-required-permissions:
      - invoice:manage
...
  /v1/invoices/{invoiceId}/receipts:
    post:
      tags:
      - Receipt
      summary: Generate invoice receipt
      description: Generate a receipt for an invoice payment using the requested terminal and template
      operationId: generateReceipt
      parameters:
      - name: invoiceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateReceiptRequest'
        required: true
      responses:
        '201':
          description: Receipt generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptResponse'
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptResponse'
      security:
      - bearerAuth: []
      x-required-permissions:
      - AUTHENTICATED
  /v1/invoices/{invoiceId}/receipts/{receiptId}/reprint:
    post:
      tags:
      - Receipt
      summary: Reprint invoice receipt
      description: Create a reprint of an existing receipt and record the reason for the reprint
      operationId: reprintReceipt
      parameters:
      - name: invoiceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: receiptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReprintReceiptRequest'
        required: true
      responses:
        '200':
          description: Receipt reprinted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptResponse'
        '404':
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptResponse'
        '409':
          description: Reprint limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptResponse'
      security:
      - bearerAuth: []
      x-required-permissions:
      - AUTHENTICATED
  /v1/invoices/{invoiceId}/receipts/{receiptId}/print:
    post:
      tags:
      - Receipt
      summary: Record printed receipt delivery
      description: Record the delivery status for a printed receipt associated with an invoice
      operationId: recordPrintDelivery
      parameters:
      - name: invoiceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: receiptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrintDeliveryRequest'
        required: true
      responses:
        '200':
          description: Print delivery recorded
        '404':
          description: Receipt not found
      security:
      - bearerAuth: []
      x-required-permissions:
      - AUTHENTICATED
  /v1/invoices/{invoiceId}/receipts/{receiptId}/email:
    post:
      tags:
      - Receipt
      summary: Email invoice receipt
      description: Send an invoice receipt by email and record the delivery status for the attempt
      operationId: sendEmailReceipt
      parameters:
      - name: invoiceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: receiptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailDeliveryRequest'
        required: true
      responses:
        '200':
          description: Email delivery recorded
        '404':
          description: Receipt not found
      security:
      - bearerAuth: []
      x-required-permissions:
      - AUTHENTICATED
...
components:
  schemas:
...
    InvoiceCreationRequest:
      type: object
      description: Internal service payload for creating invoice drafts.
      properties:
        workorderId:
          type: string
          format: uuid
          description: Workorder identifier backing the invoice draft.
          example: 550e8400-e29b-41d4-a716-446655440000
        estimateId:
          type: string
          format: uuid
          description: Source estimate identifier, when generated from estimate.
          example: 550e8400-e29b-41d4-a716-446655440001
        approvalId:
          type: string
          format: uuid
          description: Approval identifier used for billing authorization.
          example: 550e8400-e29b-41d4-a716-446655440002
        locationId:
          type: string
          format: uuid
          description: Shop location where the sale is made; used to resolve the tax jurisdiction address.
          example: 01960003-0000-7000-8000-000000000001
        idempotencyKey:
          type: string
          description: Idempotency key to prevent duplicate invoice creation.
          example: inv-create-wo-1234
        lineItems:
          type: array
          description: Line items to include on the invoice.
          items:
            $ref: '#/components/schemas/InvoiceLineItem'
    InvoiceLineItem:
      type: object
      description: Invoice line item details.
      properties:
        description:
          type: string
          description: Line item description.
          example: Brake pad replacement
        quantity:
          type: number
          description: Quantity billed.
          example: 2.0
        unitPrice:
          type: number
          description: Unit price amount.
          example: 45.0
        amount:
          type: number
          description: Line amount (quantity x unit price).
          example: 90.0
      required:
      - amount
      - description
      - quantity
      - unitPrice
    InvoiceGenerationResponse:
      type: object
      description: Response payload returned after invoice generation.
      properties:
        invoiceId:
          type: string
          format: uuid
          description: Generated invoice identifier.
          example: 550e8400-e29b-41d4-a716-446655440000
        status:
          type: string
          description: Invoice status.
          example: DRAFT
        workorderId:
          type: string
          format: uuid
          description: Source workorder identifier.
          example: 550e8400-e29b-41d4-a716-446655440001
        estimateId:
          type: string
          format: uuid
          description: Source estimate identifier.
          example: 550e8400-e29b-41d4-a716-446655440002
        approvalId:
          type: string
          format: uuid
          description: Approval identifier used for generation.
          example: 550e8400-e29b-41d4-a716-446655440003
        subtotal:
          type: number
          description: Subtotal amount before tax.
          example: 120.0
        taxAmount:
          type: number
          description: Tax amount.
          example: 9.6
        totalAmount:
          type: number
          description: Total amount after tax.
          example: 129.6
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (UTC).
          example: 2026-02-27 12:00:00+00:00
      required:
      - createdAt
      - invoiceId
      - status
      - subtotal
      - taxAmount
      - totalAmount
      - workorderId
...
    GenerateReceiptRequest:
      type: object
      description: Request to generate a receipt for an invoice payment
      properties:
        paymentIntentId:
          type: string
          format: uuid
          description: Payment intent the receipt is generated for
          example: 550e8400-e29b-41d4-a716-446655440000
        terminalId:
          type: string
          description: Identifier of the terminal producing the receipt
          example: TERM-001
          minLength: 1
        templateId:
          type: string
          description: Receipt template identifier
          example: RECEIPT_DEFAULT
          minLength: 1
        templateVersion:
          type: string
          description: Receipt template version
          example: '1'
          minLength: 1
      required:
      - paymentIntentId
      - templateId
      - templateVersion
      - terminalId
    ReceiptResponse:
      type: object
      description: Result of generating a payment receipt
      properties:
        receiptId:
          type: string
          format: uuid
          description: Unique identifier of the generated receipt
          example: 01960003-0000-7000-8000-000000000060
        reference:
          type: string
          description: Human-readable receipt reference
          example: RCPT-2026-000789
        status:
          type: string
          description: Current status of the receipt
          enum:
          - GENERATED
          example: GENERATED
      required:
      - receiptId
      - status
    ReprintReceiptRequest:
      type: object
      description: Request to reprint an existing receipt
      properties:
        reason:
          type: string
          description: Reason the receipt is being reprinted
          example: Customer requested a duplicate copy
          minLength: 1
      required:
      - reason
    PrintDeliveryRequest:
      type: object
      description: Request to record the delivery status of a printed receipt
      properties:
        status:
          type: string
          description: Print delivery outcome status
          enum:
          - SUCCESS
          - FAILED
          example: SUCCESS
      required:
      - status
    EmailDeliveryRequest:
      type: object
      description: Request to email a receipt and record the delivery status
      properties:
        emailAddress:
          type: string
          description: Recipient email address for the receipt
          example: customer@example.com
          minLength: 1
        status:
          type: string
          description: Email delivery outcome status
          enum:
          - SUCCESS
          - FAILED
          example: SUCCESS
      required:
      - emailAddress
      - status
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT